add more options and smoothing models.
Since the speckle pattern in the simulation box is a band-pass filter of the Fourier transform of the beamlets at the final focal length, it is required that m >= n. The parameters ncc and nfm are for FM SSD only. 

By comparing two ISI method, one can see the effect of different spectral shapes of phase modulation. 

One can see the reoccurrence nature of the FM SSD method.